courses/build_production_ready_agents/ch5_demos/auth/fastapi_app_auth/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-auth==2.34.0
requests==2.32.3
pydantic==2.9.0
